fea(statushandler): Refactor
